% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_graph.R
\name{categories_graph}
\alias{categories_graph}
\title{Convert Category Graph to General Graph}
\usage{
categories_graph(q, p, var_names, esti_Z, esti_Beta, categories)
}
\arguments{
\item{q}{Number of transformed variables}

\item{p}{Number of original variables}

\item{var_names}{Vector of original variable names}

\item{esti_Z}{Adjacency matrix from context-specific inference}

\item{esti_Beta}{Estimated beta coefficients for graph edges}

\item{categories}{Vector indicating the number of categories for each variable}
}
\value{
A list with:
\item{Adj_Beta}{Estimated adjacency matrix for Beta}
\item{Adj_Z}{Estimated adjacency matrix for Z}
}
\description{
This function converts a graph estimated from categorical variables
into a general adjacency matrix representation.
}
