% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_spec_graph.R
\name{context_spec_graph}
\alias{context_spec_graph}
\title{Compute Context-Specific Graph}
\usage{
context_spec_graph(q, post_Bta, post_Z, tags, bfdr)
}
\arguments{
\item{q}{Number of variables}

\item{post_Bta}{Posterior Beta samples}

\item{post_Z}{Posterior adjacency samples}

\item{tags}{Variable names}

\item{bfdr}{Bayesian False Discovery Rate}
}
\value{
Estimated graph
}
\description{
Compute Context-Specific Graph
}
