% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lambda.R
\name{find_lambda}
\alias{find_lambda}
\title{Find Optimal Lambda for Transformation}
\usage{
find_lambda(X, type)
}
\arguments{
\item{X}{Input data}

\item{type}{Vector of variable types}
}
\value{
Optimal lambda estimate
}
\description{
Find Optimal Lambda for Transformation
}
