% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_graph.R
\name{get_names_graph}
\alias{get_names_graph}
\title{Get Names for Graph}
\usage{
get_names_graph(p, q, categories, var_names)
}
\arguments{
\item{p}{Number of original variables}

\item{q}{Number of transformed variables}

\item{categories}{Category sizes}

\item{var_names}{Variable names}
}
\value{
A list of graph names and edge indicators
}
\description{
Get Names for Graph
}
