% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bmgm.R
\name{print.bmgm}
\alias{print.bmgm}
\title{Print Method for BMGM Objects}
\usage{
\method{print}{bmgm}(x, ...)
}
\arguments{
\item{x}{A bmgm object returned by \code{\link{bmgm}}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the bmgm object.
}
\description{
Print Method for BMGM Objects
}
