% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_bmgm.R
\name{sampler_bmgm}
\alias{sampler_bmgm}
\title{Gibbs Sampler for Bayesian Mixed Graphical Models}
\usage{
sampler_bmgm(
  n = 1,
  Beta,
  theta,
  type,
  categories,
  lambda,
  M = 1000,
  X_new,
  variables,
  std
)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{Beta}{Precision matrix of the graphical model.}

\item{theta}{A list of parameters for each node.}

\item{type}{Character vector indicating variable types.}

\item{categories}{Vector indicating the number of categories (for categorical variables).}

\item{lambda}{Transformation parameter.}

\item{M}{Number of Gibbs sampler iterations. Default is 1000.}

\item{X_new}{Vector of known values for imputation. Default is `c(0,0,0)`.}

\item{variables}{Indicator vector specifying which variables to update.}

\item{std}{Standard deviation used in the transformation.}
}
\value{
A matrix containing the sampled data.
}
\description{
Implements a Gibbs sampler for Bayesian inference on Mixed Graphical Models.
}
\examples{
# Simulate example data
set.seed(123)
Beta <- diag(4)  # Identity precision matrix
theta <- list(c(0,1), c(0,1), c(0.5,1), c(0.5,1))
type <- c("c", "c", "d", "d")
categories <- c(1,1,1,1)
lambda <- 1

# Run the Gibbs sampler
samples <- sampler_bmgm(n = 10, Beta = Beta, theta = theta, type = type,
                        categories = categories, lambda = lambda, M = 500)

# View first few samples
head(samples)

}
