% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_X_cat.R
\name{split_X_cat}
\alias{split_X_cat}
\title{Split Categorical Variables}
\usage{
split_X_cat(X, type)
}
\arguments{
\item{X}{Input matrix}

\item{type}{Vector of variable types}
}
\value{
A list with design matrix and category sizes
}
\description{
Split Categorical Variables
}
